Refactor : 특정 게시물 관련 커스텀훅 분리
